feat: Split router functions to allow package udpates
